requirements.txt
# Core Framework
dspy-ai>=2.4.0
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Agentic Framework
langgraph>=0.2.0
langgraph-checkpoint-postgres>=1.0.0  # PostgreSQL checkpointer for persistent state
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0  # Community integrations

# Vector Memory (Phase 0.5)
faiss-cpu>=1.7.4  # FAISS for semantic search
sentence-transformers>=2.2.0  # Required for memory embeddings
numpy>=1.24.0  # Required by FAISS

# Task Queue for Scheduled Follow-ups
celery[redis]>=5.4.0
redis>=5.0.0

# LLM Backends (switchable)
openai>=1.3.0
anthropic>=0.40.0  # Updated for latest Claude API

# Database
supabase>=2.0.0
psycopg[binary]>=3.1.0  # PostgreSQL adapter for LangGraph checkpointer
psycopg-pool>=3.1.0  # Connection pooling for PostgreSQL checkpointer

# Integrations
requests>=2.31.0
twilio>=8.10.0
sendgrid>=6.11.0
slack-sdk>=3.23.0
fastmcp>=0.1.0  # MCP client for Zapier integration (200+ tools)

# Google Drive API (for RAG pipeline)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Document Processing (for RAG pipeline)
pypdf>=3.16.2
python-docx>=1.0.1
openpyxl>=3.1.2

# Utilities
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
python-dateutil>=2.8.2

# Observability - Phoenix for LLM tracing
arize-phoenix-otel>=0.1.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
openinference-instrumentation-dspy>=0.1.0
openinference-instrumentation-langchain>=0.1.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
ruff>=0.1.6
email-validator>=2.0.0
httpx>=0.27.0  # Async HTTP client for A2A protocol
# Updated: 2025-10-26 05:39 AM - Force cache invalidation
apscheduler>=3.10.4
# Updated: 2025-10-30 15:04 - Force cache invalidation for Week 1 deployment
